+++ lib/cog_api/resources/group.ex
@@ -4,5 +4,6 @@ defmodule CogApi.Resources.Group do
   defstruct [
     :id,
     :name,
+    roles: [],
   ]
 end